requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
pinecone-client
python-dotenv
Pillow